Refactor node/registration identifier list views.
