+++ api/identifiers/views.py
@@ -1,20 +1,27 @@
+from modularodm import Q
 from rest_framework import generics, permissions as drf_permissions
 from rest_framework.exceptions import NotFound
 
 from framework.auth.oauth_scopes import CoreScopes
 
 from api.base import permissions as base_permissions
-from api.base.views import JSONAPIBaseView, BaseIdentifierList
+from api.base.views import JSONAPIBaseView
+from api.base.filters import ODMFilterMixin
 from api.base.serializers import JSONAPISerializer
 from api.base.utils import get_object_or_error
 
-from api.identifiers.serializers import IdentifierNodeSerializer, IdentifierRegistrationSerializer
+from api.identifiers.serializers import NodeIdentifierSerializer, RegistrationIdentifierSerializer
+
+from api.nodes.permissions import (
+    IsPublic,
+    ExcludeWithdrawals,
+)
 
 from website.identifiers.model import Identifier
 from website.project.model import Node
 
 
-class IdentifierList(BaseIdentifierList):
+class IdentifierList(JSONAPIBaseView, generics.ListAPIView, ODMFilterMixin):
     """List of identifiers for a specified node. *Read-only*.
 
     ##Identifier Attributes
@@ -48,7 +55,19 @@ class IdentifierList(BaseIdentifierList):
 
     """
 
-    serializer_class = IdentifierRegistrationSerializer
+    permission_classes = (
+        IsPublic,
+        drf_permissions.IsAuthenticatedOrReadOnly,
+        base_permissions.TokenHasScope,
+        ExcludeWithdrawals
+    )
+
+    required_read_scopes = [CoreScopes.IDENTIFIERS_READ]
+    required_write_scopes = [CoreScopes.NULL]
+
+    serializer_class = RegistrationIdentifierSerializer
+    node_lookup_url_kwarg = 'node_id'
+
     view_category = 'identifiers'
     view_name = 'identifier-list'
 
@@ -70,9 +89,17 @@ class IdentifierList(BaseIdentifierList):
     def get_serializer_class(self):
         if 'node_id' in self.kwargs:
             if self.get_node().is_registration:
-                return IdentifierRegistrationSerializer
-            return IdentifierNodeSerializer
+                return RegistrationIdentifierSerializer
+            return NodeIdentifierSerializer
         return JSONAPISerializer
+
+    # overrides ODMFilterMixin
+    def get_default_odm_query(self):
+        return Q('referent', 'eq', self.get_node())
+
+    # overrides ListCreateAPIView
+    def get_queryset(self):
+        return Identifier.find(self.get_query_from_request())
 
 
 class IdentifierDetail(JSONAPIBaseView, generics.RetrieveAPIView):
@@ -110,15 +137,15 @@ class IdentifierDetail(JSONAPIBaseView, generics.RetrieveAPIView):
     required_read_scopes = [CoreScopes.IDENTIFIERS_READ]
     required_write_scopes = [CoreScopes.NULL]
 
-    serializer_class = IdentifierRegistrationSerializer
+    serializer_class = RegistrationIdentifierSerializer
     view_category = 'identifiers'
     view_name = 'identifier-detail'
 
     def get_serializer_class(self):
         if 'identifier_id' in self.kwargs:
             if self.get_object().referent.is_registration:
-                return IdentifierRegistrationSerializer
-            return IdentifierNodeSerializer
+                return RegistrationIdentifierSerializer
+            return NodeIdentifierSerializer
         return JSONAPISerializer
 
     def get_object(self):
