+++ api/nodes/views.py
@@ -28,7 +28,7 @@ from api.comments.serializers import NodeCommentSerializer, CommentCreateSeriali
 from api.comments.permissions import CanCommentOrPublic
 from api.users.views import UserMixin
 from api.wikis.serializers import WikiSerializer
-from api.base.views import LinkedNodesRelationship, BaseContributorDetail, BaseContributorList, BaseNodeLinksDetail, BaseNodeLinksList, BaseLinkedList, BaseIdentifierList
+from api.base.views import LinkedNodesRelationship, BaseContributorDetail, BaseContributorList, BaseNodeLinksDetail, BaseNodeLinksList, BaseLinkedList
 from api.base.throttling import AddContributorThrottle
 
 from api.nodes.serializers import (
@@ -57,7 +57,8 @@ from api.citations.utils import render_citation
 from api.addons.serializers import NodeAddonFolderSerializer
 from api.registrations.serializers import RegistrationSerializer
 from api.institutions.serializers import InstitutionSerializer
-from api.identifiers.serializers import IdentifierNodeSerializer
+from api.identifiers.serializers import NodeIdentifierSerializer
+from api.identifiers.views import IdentifierList
 from api.nodes.permissions import (
     IsAdmin,
     IsPublic,
@@ -3178,7 +3179,7 @@ class NodeViewOnlyLinkDetail(JSONAPIBaseView, generics.RetrieveUpdateDestroyAPIV
         enqueue_postcommit_task(ban_url, (self.get_node(),), {}, celery=True, once_per_request=True)
 
 
-class NodeIdentifierList(BaseIdentifierList, NodeMixin):
+class NodeIdentifierList(NodeMixin, IdentifierList):
     """List of identifiers for a specified node. *Read-only*.
 
     ##Identifier Attributes
@@ -3212,6 +3213,4 @@ class NodeIdentifierList(BaseIdentifierList, NodeMixin):
 
     """
 
-    serializer_class = IdentifierNodeSerializer
-    view_category = 'nodes'
-    view_name = 'node-identifier-list'
+    serializer_class = NodeIdentifierSerializer
